requirements.txt
google-genai
requests
pillow
python-dotenv
